refactor(fhir): extract FhirResourceParser from FhirMediaType
